requirements-tests.txt
requests
pyone
pytest-cov
